Add files via upload
My project Hospital Management system includes registration of patients,
storing their details into the system, and also booking their appointments
with doctors. Our software has the facility to give a unique id for every
patient and stores the details of every patient and the staff automatically.
User can search availability of a doctor and the details of a patient using the
id.